refactor: move env.ts to src/ directory
