+++ packages/meilisearch/package.json
@@ -7,7 +7,7 @@
     ".": "./src/index.ts",
     "./services": "./src/services/index.ts",
     "./utils": "./src/utils/index.ts",
-    "./env": "./env.ts"
+    "./env": "./src/env.ts"
   },
   "license": "MIT",
   "scripts": {
