+++ packages/rabbit/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
   },
-  "include": ["src", "env.ts"],
+  "include": ["src"],
   "exclude": ["node_modules"]
 }
