+++ packages/s3/src/index.ts
@@ -1,5 +1,5 @@
 import { S3Client } from "@aws-sdk/client-s3"
-import { env } from "../env"
+import { env } from "./env"
 
 export const client = new S3Client({
   region: "auto",
