+++ packages/utils/src/banner.utils.ts
@@ -1,4 +1,4 @@
-import { env } from "../env"
+import { env } from "./env"
 
 const getUrl = (media: { id: string; bannerId: string }) =>
   `${env.NEXT_PUBLIC_CDN_URL}/medias/${media.id}/banners/${media.bannerId}.jpg`
