+++ packages/utils/src/cover.utils.ts
@@ -1,6 +1,6 @@
 import type { MediaCover } from "@prisma/client"
 import type { MediaCoverVolume, MediaWithRelations } from "@taiyomoe/types"
-import { env } from "../env"
+import { env } from "./env"
 
 const getUrl = (
   media:
